Added Layers to Drawing
Added Layers to Drawing the Sprites. I'm using BG1,BG2 etc. instead
of just a floating as it makes the layer more obvious to use.

Additional the View System creates an Array on every update and sorts
it. In tests so far this was faster then using spriteBatch feature of
BackToFront rendering.