Renamed Position to Transform
A Transform also takes a Direction and a Scale.

Not implemented yet, but the Direction should be used to rotate the given
Sprite if someone is assigned to it. The default direction of Down is considered
as the default. This means no Rotation is applied to a sprite.

It's down because drawing sprites can be considered drawn from up to down.

Still someone can use the View Rotation to rotate a Sprite. Both makes
sense. The Direction of a Transform is considered as the direction an
Entity looks, walks or move to. Independent of how a sprite must be rotated
to match the Look Direction.