Moved away from Texture2D and replaced it with Sprite
In general it is inconvenient to work with just Texture2D. Usually
I nearly always will have Sprite Sheets. That means I will select an
area from a bigger Texture.

And even if this is not the case and a whole Texture2D is used, then
it still makes sense to use Sprite everywhere in a Program and just
use it.

It makes the code and all the data-structures a lot easier.