Added relative positioning with Parent
Every Entity Transform can now be relative positioned to a Parent.
The Position of the Parent is inherited. When the parent has no
transform then the entity will also not rendered.

Currently Scale and Rotation is not applied. Added a box that moves over
the Knight to show the effect.