Cache Matrix
Matrix was calculated every time Camera.matrix would be called and this
function can be called very often for multiple entity without that
the camera itself changes. So i built into it that the matrix will only
calculated once.

As the camera is not fully immutable, i need to ensure to clear the current
Matrix field to `None` if some value that triggers re-calculation was
changed.