Fixed Mouse Position with RenderTarget
When Monogame uses a Viewport and for example the Game Screen has an offset
it should normaly affect the Mouse Position. But Monogame always uses the
Window Mouse Position, not the mouse position relative to the Viewport.

So this must be fixed in the Input System. So i provide the current
camera to my input system so i can subtract the Viewport information.