Fixed null reference exception & Input Handling
Moved the Camera definition out of the Input definition. As the camera
can only be available when the game already runs. But the input definition
should be defined outside the running game. As it anyway should never
change, or at least not at every frame. So it also don't need to be build
every frame.

On top, forget to set `resetInput` back to `false` once it was resetted.