Nicer Animation of Boxes
All 3000 boxes now use a SpriteSheetAnimation that loops through
an animation. On top the move constantly at fixedUpdate interval in
a random direction. This tests Movement Component.

Performance at this point stays nearly unchanged. In this version
I also added that all 3,000 boxes now rotate around an origin point.

Assigning a parent cost much performance because a transform matrix
of parent position, rotation and scale must be computed. This drops
performance visible. from 2200 fps to 1300 fps. But still good enough
to not care much about it. Maybe can still be optimized with some caching.

Tryed a caching but was not succesfull. Reason is that having just one
parent has nearly no performance impact. Caching the result
of matrix makes only sense if we have a deep hirachy of many parents
that have other parents.

But in this "game engine" this is usually not what someone needs todo
very often. And when they do they likely only have one parent and
very few childs, so makes less sense to add caching at this point.