Expanded Movement
During testing I encountered some problems and thought some of the
calculation of the parent was wrong. Took me some time to fix everything
and also thought about what was right/wrong.

Changed the System in such a way that now Vector2.Up is considered
the default direction. Or in that case Up Direction. All rotation are
relative to up. So a 90° degree rotation points to the right.

Local movement to the right now also means something moves to the right.
It works nicely for 2d top down games but also for 2d side-scrollers.

But it changed the way how vectors to angles and in reverse are computed.