Changed Transform
It now saves a Rotation as radiant instead of a Vector2. Radiant is
anyway the most used way how I use rotation instead of using a Vector2.
When the Transform was computed it also was converted from Vector2
to radiant. Now it is always Radiant what should reduce computation
somewhat.

Otherwise added some inline statement and some comments throughout the code.