Performance improvement
I anyway will change this later, but this change already improvements
performance a lot when a lot of boxes are not showing. Instead of
having an array with all transform,view i just build an array of those
that are visible.

So when a lot are not showing, like in the current code. it will only build
an array with 1800 transform,view in it. Instead of all 3600+ entities.

Sorting 1800 elements is faster than 3600+, it also omits a fetch of getting
a transform for those elements that are not showing. And in the draw loop
i only iterate those who are showing without the need of an if-statement.

Greatly improves performance when a lot of boxes are not showing.