Removed View.IsVisible
Instead of a boolean that contains the information if something is visible
and thus should be rendered, i built up two dictionaries, one containing
the visible entities and the other the hidden elements.

Changing the visibility comes at a higher cost because the view has to be
removed/add to different containers. The pure rendering becomes faster
because it just can iterate through the visible entities and it has no
branch (if-statement) anymore.

Having lots of entities that also contains lot of hidden elements is now
overall of better performance.