Changed Animation
Animation was poorly done before. It mixed the definition of Animation
and is sheets with the current selected sprite. This was the reason why
I needed copy before.

Changed it in the way that now Sheets is an immutable data-structure. Once
loaded it anyway never should change and the Sheets could be shared
between multiple instances.

Animation now contains the sheets and besides that all data for which
sheet and sprite should be shown. Animation is the Component for every
Entity.

Also changed it in a way that the current Sprite is mutable and animation
just updates the current View instead of creating a new View every time
the sprite changes.

This also increased performance a little bit, but mostly the best benefit
should be less garbage generation and GC time.