Changed Camera.origin // Added comments/type-information
Now holds a Vector3 for Origin as this is anyway needed when the matrix
is created. This value usually never changes once set so computing
it on every frame was somehow pointless.

Otherwise added some comment and type-information. Hopefully this way
compilation becomes a little bit faster. But sometimes also can help
to see faster what needs to be passed or returned.