Removed Caching system
Is not needed anymore. Before I had a Cache that cached all Entities
that appear in multiple collections. For example in Transform and View.
So i could only iterate through those Entities that had both. But
then I still needed to fetch the Transform and View Component. At least
in that code fetching never failed.

Now i just iterate through the Dictionary directly. For example in the
View System i start iterating through the View dictionary. This gives me
the Entity and the View. Then from there on i fetch the Transform.

Now this can fail so probably can lead to branch mispredictions, but
this reduces the overhead of a Cache, an Event that fires whenever Component
gets added/removed and makes code in general easier.

Maybe in the future i could add information to the Entity that stores
which Component it has as a BitField. But it didn't avoid the need to
fetch the Component from a Dictionary. So also could be more overhead,
depends on the data/uses if it makes it faster or not.