Removed Transform.empty
As Transform is mutable this is not a good idea. It can create a static
property Empty and only creates the structure once. But when some code
changes Empty all other kind of initialization becomes wrong.

Depends on how it is exactly implemented. Also could be that it turns
out to be a Property that creates it, but I don't want to rely on
some F# compiler implementation details.